Version 2 of the GS
This version has working tables and main graph. Data in is parsed to the proper table (payload or on-person device), and the main graph can plot all the telemetry of either input. Not sure how two radio inputs will work, but we'll see. Still to be done is getting EEG data in (its a work in progress) and the other plot (for proximity data I believe).